+++ packages/standards/src/application/adapter/StandardsAdapter.ts
@@ -50,26 +50,26 @@ import { IStandardsRepositories } from '../../domain/repositories/IStandardsRepo
 import { GetRuleExamplesCommand } from '../../domain/useCases';
 import { GenerateStandardSummaryJobFactory } from '../../infra/jobs/GenerateStandardSummaryJobFactory';
 import { StandardsServices } from '../services/StandardsServices';
-import { AddRuleToStandardUsecase } from '../useCases/addRuleToStandard/addRuleToStandard.usecase';
-import { CreateRuleExampleUsecase } from '../useCases/createRuleExample/createRuleExample.usecase';
-import { CreateStandardUsecase } from '../useCases/createStandard/createStandard.usecase';
-import { CreateStandardSamplesUsecase } from '../useCases/createStandardSamples/createStandardSamples.usecase';
-import { CreateStandardWithExamplesUsecase } from '../useCases/createStandardWithExamples/createStandardWithExamples.usecase';
-import { CreateStandardWithPackagesUsecase } from '../useCases/createStandardWithPackages/createStandardWithPackages.usecase';
-import { DeleteRuleExampleUsecase } from '../useCases/deleteRuleExample/deleteRuleExample.usecase';
-import { DeleteStandardUsecase } from '../useCases/deleteStandard/deleteStandard.usecase';
-import { DeleteStandardsBatchUsecase } from '../useCases/deleteStandardsBatch/deleteStandardsBatch.usecase';
-import { FindStandardBySlugUsecase } from '../useCases/findStandardBySlug/findStandardBySlug.usecase';
-import { GetLatestStandardVersionUsecase } from '../useCases/getLatestStandardVersion/getLatestStandardVersion.usecase';
-import { GetRuleExamplesUsecase } from '../useCases/getRuleExamples/getRuleExamples.usecase';
-import { GetRulesByStandardIdUsecase } from '../useCases/getRulesByStandardId/getRulesByStandardId.usecase';
-import { GetStandardByIdUsecase } from '../useCases/getStandardById/getStandardById.usecase';
-import { GetStandardVersionUsecase } from '../useCases/getStandardVersion/getStandardVersion.usecase';
-import { GetStandardVersionByIdUsecase } from '../useCases/getStandardVersionById/getStandardVersionById.usecase';
-import { ListStandardsBySpaceUsecase } from '../useCases/listStandardsBySpace/listStandardsBySpace.usecase';
-import { ListStandardVersionsUsecase } from '../useCases/listStandardVersions/listStandardVersions.usecase';
-import { UpdateRuleExampleUsecase } from '../useCases/updateRuleExample/updateRuleExample.usecase';
-import { UpdateStandardUsecase } from '../useCases/updateStandard/updateStandard.usecase';
+import { AddRuleToStandardUsecase } from '../useCases/addRuleToStandard/AddRuleToStandardUsecase';
+import { CreateRuleExampleUsecase } from '../useCases/createRuleExample/CreateRuleExampleUsecase';
+import { CreateStandardUsecase } from '../useCases/createStandard/CreateStandardUsecase';
+import { CreateStandardSamplesUsecase } from '../useCases/createStandardSamples/CreateStandardSamplesUsecase';
+import { CreateStandardWithExamplesUsecase } from '../useCases/createStandardWithExamples/CreateStandardWithExamplesUsecase';
+import { CreateStandardWithPackagesUsecase } from '../useCases/createStandardWithPackages/CreateStandardWithPackagesUsecase';
+import { DeleteRuleExampleUsecase } from '../useCases/deleteRuleExample/DeleteRuleExampleUsecase';
+import { DeleteStandardUsecase } from '../useCases/deleteStandard/DeleteStandardUsecase';
+import { DeleteStandardsBatchUsecase } from '../useCases/deleteStandardsBatch/DeleteStandardsBatchUsecase';
+import { FindStandardBySlugUsecase } from '../useCases/findStandardBySlug/FindStandardBySlugUsecase';
+import { GetLatestStandardVersionUsecase } from '../useCases/getLatestStandardVersion/GetLatestStandardVersionUsecase';
+import { GetRuleExamplesUsecase } from '../useCases/getRuleExamples/GetRuleExamplesUsecase';
+import { GetRulesByStandardIdUsecase } from '../useCases/getRulesByStandardId/GetRulesByStandardIdUsecase';
+import { GetStandardByIdUsecase } from '../useCases/getStandardById/GetStandardByIdUsecase';
+import { GetStandardVersionUsecase } from '../useCases/getStandardVersion/GetStandardVersionUsecase';
+import { GetStandardVersionByIdUsecase } from '../useCases/getStandardVersionById/GetStandardVersionByIdUsecase';
+import { ListStandardsBySpaceUsecase } from '../useCases/listStandardsBySpace/ListStandardsBySpaceUsecase';
+import { ListStandardVersionsUsecase } from '../useCases/listStandardVersions/ListStandardVersionsUsecase';
+import { UpdateRuleExampleUsecase } from '../useCases/updateRuleExample/UpdateRuleExampleUsecase';
+import { UpdateStandardUsecase } from '../useCases/updateStandard/UpdateStandardUsecase';
 
 const origin = 'StandardsAdapter';
