+++ packages/standards/src/application/useCases/deleteRuleExample/DeleteRuleExampleUsecase.ts
