+++ packages/standards/src/application/useCases/getStandardById/GetStandardByIdUsecase.spec.ts
@@ -16,7 +16,7 @@ import { v4 as uuidv4 } from 'uuid';
 import { standardFactory } from '../../../../test/standardFactory';
 import { createStandardId } from '@packmind/types';
 import { StandardService } from '../../services/StandardService';
-import { GetStandardByIdUsecase } from './getStandardById.usecase';
+import { GetStandardByIdUsecase } from './GetStandardByIdUsecase';
 
 describe('GetStandardByIdUsecase', () => {
   let usecase: GetStandardByIdUsecase;
